fix(sync): record the accepted page size, and split the retry helpers out
Review follow-ups on the page-size retry, plus one build break the retry commit
carried in.

- **`page_size_override` records the size the provider accepted**, not the one
  the shrink last tried. With several halvings (25 -> 12 -> 6) the old
  assignment left the previous *rejected* size in the override on every step but
  the last, and was right at the end only because that step happened to be the
  accepted one. The value is now set where the response succeeds, so the
  "sticky for the run" intent no longer depends on the retry order.
- **`413` has to stand alone.** An unanchored `contains("413")` also matches a
  message id, an amount, or a timestamp carrying those digits, and each false
  positive costs a shrink-and-retry cycle before the real error surfaces — on a
  failure a smaller page was never going to fix.
- **The page-size helpers move to `page_size.rs`** with their own tests, putting
  `orchestrator.rs` back under the 500-line cap.
- **`tracing` is gated behind `sync`.** It is an optional dependency, but the
  chunk store's new cold-start retry called it unconditionally, so the
  dependency-light default build could not compile. CI builds with the feature
  on, which is why it stayed green.
- One `cloned_ref_to_slice_refs` lint in a test this PR touches.

Tests: 1301 lib tests green with `--features sync`; the default-feature build
compiles; clippy clean in both configurations.
